Rolls code generation into proto generation rules and scripts for FhirPackages. This means a single gen proto rule with a single FhirPackage will generate all codes, profiles, and extensions.
PiperOrigin-RevId: 280134441

+++ bazel/generate_protos_utils.sh
@@ -26,7 +26,7 @@ function copy_to_src_if_present() {
   src=bazel-genfiles/$dir/_genfiles_$1
 
   dst=$dir/$1
-  if [ -e $src ]
+  if [ -e $src ] && [ $(wc -l < $src) -gt 20 ]
   then
     cp $src $dst
   fi
